Entradas de dados
Aqui aprendemos entrada de dados
Scanner:
next()
nextInt()
nextDouble()
next().charAt(0)

Locale

Como ler até a quebra de linha
nextLine()
como limpar o buffer de leitura

+++ TesteProcessamento/TesteProcessamento.java
@@ -7,7 +7,7 @@ public class TesteProcessamento {
         a = 5;
         b = 2;
 
-        resultado = a/b;
+        resultado = (double)a/b;
 
         System.out.println(resultado);
     }
